// src/app/_core/models/User.model.ts
export interface User {
    id: number;
    // name: string;
    // phoneNumber: string;
    // password: string;
    // passwordConfirmation: string;
    role: number;
    // idAgence: number;
    // firstname: string;
    // lastname: string;
    username: string;
    // phone: string;
    email: string;
    password: string;
    balance: number;
    
    // notificationToken: string;
}